# docs/swagger.yaml
basePath: /api/
definitions:
  models.Content:
    properties:
      content_size:
        type: integer
      content_type:
        type: string
    type: object
  models.S3FileInfo:
    properties:
      content:
        $ref: '#/definitions/models.Content'
      file_id:
        type: string
      file_key:
        type: string
      file_name:
        type: string
    type: object
  models.S3FileTask:
    properties:
      completed_at:
        type: string
      created_at:
        type: string
      download_url:
        type: string
      error:
        type: string
      file_info:
        $ref: '#/definitions/models.S3FileInfo'
      id:
        type: string
      processed_key:
        type: string
      status:
        $ref: '#/definitions/models.TaskStatus'
      updated_at:
        type: string
    type: object
  models.TaskStatus:
    enum:
    - pending
    - processing
    - completed
    - failed
    type: string
    x-enum-varnames:
    - TaskStatusPending
    - TaskStatusProcessing
    - TaskStatusCompleted
    - TaskStatusFailed
  models.UploadResponse:
    properties:
      key:
        type: string
      message:
        type: string
      size:
        type: integer
      task_id:
        type: string
      task_status:
        type: string
      url:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API для обработки изображений
  title: Image Sketch Processor API
  version: "1.0"
paths:
  /files:
    post:
      consumes:
      - multipart/form-data
      description: Загружает изображение в S3 и создает задачу на .
      parameters:
      - description: Изображение (JPG, PNG, max 10MB)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Task создана, файл в S3
          schema:
            $ref: '#/definitions/models.UploadResponse'
        "400":
          description: Неверный файл
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать задачу на обработку изображения
      tags:
      - files
  /tasks/{id}:
    get:
      description: Получить текущий статус задачи по ID
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Задача
          schema:
            $ref: '#/definitions/models.S3FileTask'
        "404":
          description: Задача не найдена
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить статус обработки файла
      tags:
      - tasks
swagger: "2.0"